Fixed importing from OpenStreetMaps
